+++ NEWS.md
@@ -1,3 +1,5 @@
+# ggstats (development version)
+
 # ggstats 0.7.0
 
 **Minor breaking change**